UI Improvement: Create Request
