+++ package.json
@@ -21,7 +21,6 @@
     "disqus-react": "^1.0.11",
     "firebase": "^8.4.1",
     "lodash": "^4.17.21",
-    "material-ui-phone-number": "^2.2.6",
     "moment": "^2.29.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -29,7 +28,6 @@
     "react-jss": "^10.6.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
-    "react-select": "^4.3.0",
     "styled-components": "^5.2.3",
     "typescript": "^4.2.4",
     "web-vitals": "^1.1.1",
