+++ src/global.d.ts
@@ -1,0 +1,6 @@
+declare module '*.jpg' {
+    export default '' as string;
+  }
+  declare module '*.png' {
+    export default '' as string;
+  }
